add default circle styles
