+++ packages/lesmis/src/components/ListVisual/index.js
@@ -1,0 +1,6 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+export * from './ListVisual';
