+++ packages/ui-components/.storybook/moduleMock.js
@@ -1,0 +1,5 @@
+// a mock module for replacing modules in other packages, such as `yargs` within `@tupaia/utils`,
+// that breaks storybook
+module.exports = {
+  strict: () => {},
+};
