--- packages/ui-components/.storybook/yargsMock.js
@@ -1,5 +1,0 @@
-// mock implementation of `yargs`, because `script.js` within `@tupaia/utils` uses it but that
-// breaks storybook
-module.exports = {
-  strict: () => {},
-};
